refactor: Rename parser to metainfo
