+++ crates/libtortillas/src/lib.rs
@@ -1,6 +1,6 @@
 pub mod engine;
 pub mod errors;
 pub mod hashes;
-pub mod parser;
+pub mod metainfo;
 pub mod peers;
 pub mod tracker;
