+++ crates/libtortillas/src/tracker/http.rs
@@ -506,7 +506,7 @@ mod tests {
 
    use super::HttpTracker;
    use crate::{
-      parser::{MetaInfo, TorrentFile},
+      metainfo::{MetaInfo, TorrentFile},
       tracker::TrackerTrait,
    };
